// src/app/core/models/masters/mrm/meeting-category.ts
export interface MeetingCategory {
    id: number;
    title: string;
    description:string;
    status: string;
    status_id :number;
    status_label: string;
}

export interface MeetingCategoryPaginationResponse {
    current_page: number;
    from:number;
    total: number;
    per_page: number;
    last_page: number;
    data: MeetingCategory[];
}